backend/requirements.txt
# FastAPI Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.39
alembic==1.13.1
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Security & Passwords
passlib==1.7.4
PyJWT==2.12.1
bcrypt==4.0.1

# Environment Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.0.0
python-multipart==0.0.6
aiosmtplib==3.0.2

# Development (optional)
requests==2.32.3

# Utilities
python-dotenv==1.0.0
# Testing (optional)
pytest==7.4.3
httpx==0.25.2
